+++ src/Ext/Techno/Body.cpp
@@ -1,4 +1,4 @@
-#include "Body.h"
+﻿#include "Body.h"
 
 #include <Ext/Aircraft/Body.h>
 #include <Ext/Anim/Body.h>
@@ -495,38 +495,50 @@ bool TechnoExt::HasAttachedEffects(std::vector<AttachEffectTypeClass*> attachEff
 
 	for (auto const& type : attachEffectTypes)
 	{
-		for (auto const& attachEffect : this->AttachedEffects)
+		if (type->Cumulative)
 		{
-			if (attachEffect->GetType() == type && attachEffect->IsActive())
+			const int cumulativeCount = this->GetAttachedEffectCumulativeCount(type, ignoreSameSource, pInvoker, pSource);
+			bool matched = cumulativeCount > 0;
+			const unsigned int minSize = minCounts ? minCounts->size() : 0;
+			const unsigned int maxSize = maxCounts ? maxCounts->size() : 0;
+
+			if (matched && minSize > 0)
 			{
-				if (checkSource && attachEffect->IsFromSource(pInvoker, pSource))
-					continue;
+				if (cumulativeCount < minCounts->at(typeCounter - 1 >= minSize ? minSize - 1 : typeCounter - 1))
+					matched = false;
+			}
 
-				const unsigned int minSize = minCounts ? minCounts->size() : 0;
-				const unsigned int maxSize = maxCounts ? maxCounts->size() : 0;
+			if (matched && maxSize > 0)
+			{
+				if (cumulativeCount > maxCounts->at(typeCounter - 1 >= maxSize ? maxSize - 1 : typeCounter - 1))
+					matched = false;
+			}
 
-				if (type->Cumulative && (minSize > 0 || maxSize > 0))
-				{
-					const int cumulativeCount = this->GetAttachedEffectCumulativeCount(type, ignoreSameSource, pInvoker, pSource);
-
-					if (minSize > 0)
-					{
-						if (cumulativeCount < minCounts->at(typeCounter - 1 >= minSize ? minSize - 1 : typeCounter - 1))
-							continue;
-					}
-					if (maxSize > 0)
-					{
-						if (cumulativeCount > maxCounts->at(typeCounter - 1 >= maxSize ? maxSize - 1 : typeCounter - 1))
-							continue;
-					}
-				}
-
+			if (matched)
+			{
 				// Only need to find one match, can stop here.
 				if (!requireAll)
 					return true;
 
 				foundCount++;
-				break;
+			}
+		}
+		else
+		{
+			for (auto const& attachEffect : this->AttachedEffects)
+			{
+				if (attachEffect->GetType() == type && attachEffect->IsActive())
+				{
+					if (checkSource && attachEffect->IsFromSource(pInvoker, pSource))
+						continue;
+
+					// Only need to find one match, can stop here.
+					if (!requireAll)
+						return true;
+
+					foundCount++;
+					break;
+				}
 			}
 		}
 
@@ -553,9 +565,6 @@ bool TechnoExt::HasAttachedEffects(std::vector<AttachEffectTypeClass*> attachEff
 /// <returns>Number of active cumulative AttachEffect type instances on the techno. 0 if the AttachEffect type is not cumulative.</returns>
 int TechnoExt::GetAttachedEffectCumulativeCount(AttachEffectTypeClass* pAttachEffectType, bool ignoreSameSource, TechnoClass* pInvoker, AbstractClass* pSource) const
 {
-	if (!pAttachEffectType->Cumulative)
-		return 0;
-
 	unsigned int foundCount = 0;
 	const bool checkSource = ignoreSameSource && pInvoker && pSource;
 
@@ -590,9 +599,11 @@ void TechnoExt::CreateDelayedFireAnim(TechnoClass* pThis, AnimTypeClass* pAnimTy
 {
 	if (pAnimType)
 	{
-		auto coords = pThis->GetCenterCoords();
+		CoordStruct coords;
 
-		if (!center)
+		if (center)
+			coords = pThis->GetCenterCoords();
+		else
 			coords = TechnoExt::GetFLHAbsoluteCoords(pThis, firingCoords, onTurret);
 
 		auto const pAnim = GameCreate<AnimClass>(pAnimType, coords);
@@ -993,14 +1004,33 @@ bool __fastcall TechnoExt::ApplyKillDriver(TechnoClass** pData, void*, HouseClas
 int TechnoExt::GetSight()
 {
 	double sight = this->TypeExtData->OwnerObject()->Sight;
-	
+
 	for (auto& callback : TechnoExtInterop::CalculateSightCallbacks)
 	{
 		if (callback)
 			sight = callback(this->OwnerObject(), sight);
 	}
-	
+
 	return static_cast<int>(sight);
+}
+
+bool TechnoExt::CanReceiveEvent(TechnoClass* pThis, HouseClass* pHouse)
+{
+	if (pThis->Berzerk)
+		return false;
+
+	if (pThis->GetTechnoType()->Spawned)
+		return false;
+
+	if (pThis->SlaveOwner)
+		return false;
+
+	auto const pOwner = pThis->GetOwningHouse();
+
+	if (pOwner != pHouse && !(pHouse->IsCurrentPlayer() && pOwner->IsControlledByCurrentPlayer()))
+		return false;
+
+	return true;
 }
 
 bool TechnoExt::HasWeaponsDisabled(TechnoClass* pThis)