+++ src/New/Entity/AttachEffectClass.cpp
@@ -150,6 +150,9 @@ void AttachEffectClass::PointerGotInvalid(void* ptr, bool removed)
 				{
 					AnnounceInvalidPointer(pEffect->Invoker, ptr);
 
+					if ((pEffect->Type->DiscardOn & DiscardCondition::InvokerDie) != DiscardCondition::None)
+						pEffect->ShouldBeDiscarded = true;
+
 					if (--count == 0)
 						break;
 				}
@@ -453,18 +456,12 @@ void AttachEffectClass::KillAnim()
 	}
 }
 
-void AttachEffectClass::UpdateCumulativeAnim()
+void AttachEffectClass::UpdateCumulativeAnim(int count)
 {
-	if (!this->HasCumulativeAnim)
-		return;
-
 	const auto pAnim = this->Animation;
 
 	if (!pAnim)
 		return;
-
-	const auto pType = this->Type;
-	const int count = TechnoExt::Fetch(this->Techno)->GetAttachedEffectCumulativeCount(pType);
 
 	if (count < 1)
 	{
@@ -472,22 +469,11 @@ void AttachEffectClass::UpdateCumulativeAnim()
 		return;
 	}
 
+	const auto pType = this->Type;
 	auto const pAnimType = pType->GetCumulativeAnimation(count);
 
 	if (pAnim->Type != pAnimType)
 		AnimExt::ChangeAnimType(pAnim, pAnimType, false, pType->CumulativeAnimations_RestartOnChange);
-}
-
-void AttachEffectClass::TransferCumulativeAnim(AttachEffectClass* pSource)
-{
-	if (!pSource || !pSource->Animation)
-		return;
-
-	this->KillAnim();
-	this->Animation = pSource->Animation;
-	this->HasCumulativeAnim = true;
-	pSource->Animation = nullptr;
-	pSource->HasCumulativeAnim = false;
 }
 
 void AttachEffectClass::SetAnimationTunnelState(bool visible)
@@ -703,6 +689,7 @@ int AttachEffectClass::Attach(TechnoClass* pTarget, HouseClass* pInvokerHouse, T
 	bool decloak = false;
 	double ROFModifier = 1.0;
 	const bool selfOwned = pTarget == pSource;
+	std::set<AttachEffectTypeClass*> cumulativeAnimTypes;
 
 	for (size_t i = 0; i < types.size(); i++)
 	{
@@ -725,7 +712,7 @@ int AttachEffectClass::Attach(TechnoClass* pTarget, HouseClass* pInvokerHouse, T
 					markForRedraw = true;
 
 				if (pType->Cumulative && pType->CumulativeAnimations.size() > 0)
-					pTargetExt->UpdateCumulativeAttachEffects(pType);
+					cumulativeAnimTypes.insert(pType);
 			}
 		}
 	}
@@ -748,6 +735,11 @@ int AttachEffectClass::Attach(TechnoClass* pTarget, HouseClass* pInvokerHouse, T
 
 		if (decloak && pTarget->CloakState == CloakState::Cloaked)
 			pTarget->Uncloak(true);
+	}
+
+	for (auto const pType : cumulativeAnimTypes)
+	{
+		pTargetExt->UpdateCumulativeAttachEffects(pType);
 	}
 	          
 	return attachedCount;
@@ -982,7 +974,8 @@ int AttachEffectClass::RemoveAllOfType(AttachEffectTypeClass* pType, TechnoClass
 
 	auto const targetAEs = &pTargetExt->AttachedEffects;
 	std::vector<std::unique_ptr<AttachEffectClass>>::iterator it;
-	std::vector<std::pair<WeaponTypeClass*, TechnoClass*>> expireWeapons;
+	std::vector<AEWeaponParams> expireWeapons;
+	std::set<AttachEffectTypeClass*> cumulativeAnimTypes;
 
 	for (it = targetAEs->begin(); it != targetAEs->end(); )
 	{
@@ -1003,17 +996,19 @@ int AttachEffectClass::RemoveAllOfType(AttachEffectTypeClass* pType, TechnoClass
 					if (pType->ExpireWeapon_UseInvokerAsOwner)
 					{
 						if (auto const pInvoker = attachEffect->Invoker)
-							expireWeapons.push_back(std::make_pair(pType->ExpireWeapon, pInvoker));
+							expireWeapons.push_back(AEWeaponParams { pType->ExpireWeapon, pInvoker, pInvoker->Owner });
+						else
+							expireWeapons.push_back(AEWeaponParams { pType->ExpireWeapon, nullptr, attachEffect->GetInvokerHouse() });
 					}
 					else
 					{
-						expireWeapons.push_back(std::make_pair(pType->ExpireWeapon, pTarget));
+						expireWeapons.push_back(AEWeaponParams { pType->ExpireWeapon, pTarget, pTarget->Owner });
 					}
 				}
 			}
 
 			if (pType->Cumulative && pType->CumulativeAnimations.size() > 0)
-				pTargetExt->UpdateCumulativeAttachEffects(pType, attachEffect);
+				cumulativeAnimTypes.insert(pType);
 
 			if (attachEffect->ResetIfRecreatable())
 			{
@@ -1034,12 +1029,16 @@ int AttachEffectClass::RemoveAllOfType(AttachEffectTypeClass* pType, TechnoClass
 		}
 	}
 
+	for (auto const type : cumulativeAnimTypes)
+	{
+		pTargetExt->UpdateCumulativeAttachEffects(type, true);
+	}
+
 	auto const coords = pTarget->GetCoords();
 
-	for (auto const& pair : expireWeapons)
+	for (auto const& info : expireWeapons)
 	{
-		auto const pInvoker = pair.second;
-		WeaponTypeExt::DetonateAt(pair.first, coords, pInvoker, pInvoker->Owner, pTarget);
+		WeaponTypeExt::DetonateAt(info.Weapon, coords, info.Invoker, info.InvokerHouse, pTarget);
 	}
 
 	return detachedCount;