+++ src/New/Entity/AttachEffectClass.h
@@ -20,13 +20,12 @@ public:
 	void AI_Temporal();
 	void KillAnim();
 	void CreateAnim();
-	void UpdateCumulativeAnim();
-	void TransferCumulativeAnim(AttachEffectClass* pSource);
+	void UpdateCumulativeAnim(int count);
 
-	bool CanShowAnim(bool skipAnimCheck = false) const
+	bool CanShowAnim() const
 	{
 		const auto pType = this->Type;
-		return (pType->Animation || skipAnimCheck || this->HasCumulativeAnim)
+		return (pType->Animation || (pType->Cumulative && pType->CumulativeAnimations.size() > 0))
 			&& (this->IsOnline || pType->Animation_OfflineAction != AttachedAnimFlag::Hides)
 			&& (!this->IsUnderTemporal || pType->Animation_TemporalAction != AttachedAnimFlag::Hides)
 			&& !this->IsAnimHidden && !this->IsInTunnel;
@@ -42,6 +41,7 @@ public:
 	bool ShouldBeDiscardedNow();
 	bool IsFromSource(TechnoClass* pInvoker, AbstractClass* pSource) const { return pInvoker == this->Invoker && pSource == this->Source; }
 	TechnoClass* GetInvoker() const { return this->Invoker; }
+	HouseClass* GetInvokerHouse() const { return this->InvokerHouse; }
 	bool IsActive() const { return this->IsOnline && this->IsActiveIgnorePowered(); }
 
 	bool IsActiveIgnorePowered() const